Removed 'options' argument.
Options may be needed in the future, but not any soon. Boutique is going to be used with sensible defaults and won't be parametrized by Apiary app or by user in the near future. Thus after some thinking I assumed the 'options' argument is redundant at the moment and unnecessarily pollutes interfaces of many functions. Once it is actually needed and there is a real use case for it, it can be easily re-introduced.